requirements.txt
Flask
replicate
python-dotenv
Pillow